+++ test/productsById.test.js
@@ -1,5 +1,5 @@
 import { handler } from "../src/productsById";
-import ProductRepository from "../src/productModule/productRepository";
+import ProductRepository from "../src/productService/productRepository";
 
 const mockedProduct = { id: "1ac5310f-1025-4196-b49e-5188d1f0f63c" };
 
@@ -7,7 +7,7 @@ jest.mock("../src/utils/envUtils", () => ({
   checkRequiredEnvValues: jest.fn(),
 }));
 
-jest.mock("../src/productModule/productRepository");
+jest.mock("../src/productService/productRepository");
 
 describe("productById", () => {
   beforeAll(() => {
@@ -35,7 +35,7 @@ describe("productById", () => {
   });
 
   it("returns 422 with message if parameter is not support uuid format", async () => {
-    const { headers, statusCode, body } = await handler({ pathParameters: { productId: 25} });
+    const { headers, statusCode, body } = await handler({ pathParameters: { productId: 25 } });
 
     expect(headers).toBeDefined();
     expect(statusCode).toBe(422);
